feat: 컴포넌트 병합(불필요 컴포넌트 제거)
